treqsy_backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
fastapi[all] 